Adding on_settings_migrate() to add the 2 new default options to older configurations.  This should fix all the KeyErrors people are seeing.
Set the pigpiod pwm dutycycle to range from 0-100 so that it matches
RPi.GPIO.

The startup pwm value should be 0 or 100. This should fix the
ValueError users were seeing.

Removed some of the debugging cruft I had added trying to fix the
original bugs, but I didn't have an actual printer around to test
with.